// superglue/lib/types/page.ts
export type RestoreStrategy =
  | 'fromCacheOnly'
  | 'revisitOnly'
  | 'fromCacheAndRevisitInBackground'

/**
 * A NavigationAction is used to tell Superglue to history.push, history.replace
 * or do nothing.
 */
export type NavigationAction = 'push' | 'replace' | 'none'

/**
 * An identifier that Superglue will uses to determine which page component to render
 * with your page response.
 */
export type ComponentIdentifier = string

/**
 * A keypath is a string representing the location of a piece of data. Superglue uses
 * the keypath to dig for or update data.
 *
 * @example
 * Object access
 * ```
 * data.header.avatar
 * ```
 *
 * @example
 * Array access
 * ```
 * data.body.posts.0.title
 * ```
 *
 * @example
 * Array with lookahead
 * ```
 * data.body.posts.post_id=foobar.title
 * ```
 */
export type Keypath = string

export type FlashState = Record<string, JSONValue>
// ...
export type Defer = {
  url: string
  type: 'auto' | 'manual'
  path: Keypath
  successAction: string
  failAction: string
}
// ...
export type FragmentPath = {
  id: string
  path: Keypath
}

/**
 * The SaveResponse response is responsible for persisting a full page
 * visit in Superglue.
 */
export type SaveResponse<T = JSONMappable> = {
  data: T
  componentIdentifier: ComponentIdentifier
  assets: string[]
  csrfToken?: string
  fragments: FragmentPath[]
  defers: Defer[]
  flash: FlashState
  action: 'savePage'

  renderedAt: number
  restoreStrategy: RestoreStrategy
}

/**
 * A Page is a SaveResponse that's been saved to the store
 */
export type Page<T = JSONMappable> = SaveResponse<T> & {
  savedAt: number
}

/**
 * The GraftResponse is responsible for partial updates using props_template's
 * digging functionality in Superglue.
 *
 * @property path Used by superglue to replace the data at that location.
 * @property equals to `graft` to indicate a {@link GraftResponse}
 * @interface
 */
export type GraftResponse<T = JSONMappable> = {
  data: T
  componentIdentifier: ComponentIdentifier
  assets: string[]
  csrfToken?: string
  fragments: FragmentPath[]
  defers: Defer[]
  flash: FlashState
  action: 'graft'
  renderedAt: number

  path: Keypath
  fragmentContext?: string
}